requirements.txt
asgiref==3.8.1
chardet==5.2.0
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.58.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
